sealing: Drop redundant Piece type
